+++ src/app/(dashboard)/(pages)/pages/components/list/filters/filters.tsx
@@ -1,0 +1,11 @@
+import ListFilters from '@/components/list/filters/filters';
+
+import PagesListFilterStatus from './status';
+
+export default function PagesListFilters() {
+  return (
+    <ListFilters>
+      <PagesListFilterStatus />
+    </ListFilters>
+  );
+}