+++ src/app/(dashboard)/(pages)/pages/components/page/wrapper.tsx
@@ -1,0 +1,31 @@
+'use client';
+
+import DeleteDialog from '@/components/dialogs/delete';
+
+import { usePagesStore } from '../../store/pages-store';
+
+export default function PagesPageWrapper({ children }: { children: React.ReactNode }) {
+  const { deleteDialogOpen, setDeleteDialogOpen, deleteDialogPageId } = usePagesStore();
+
+  const deleteDialogClose = () => {
+    setDeleteDialogOpen(false);
+  };
+
+  return (
+    <>
+      {children}
+
+      {/* delete page dialog */}
+      <DeleteDialog
+        dialogOpen={deleteDialogOpen}
+        dialogClose={deleteDialogClose}
+        deleteRowId={deleteDialogPageId}
+        deleteTable="pages"
+        deletePath="/pages"
+        deleteLogEventName="DELETE_PAGE"
+        deleteLogEventMessage="Page"
+        title="Delete page"
+      />
+    </>
+  );
+}